+++ ui/src/logic/useSystemUpdate.tsx
@@ -12,9 +12,9 @@ function pikeIsBlocked(newKelvin: number, pike: Pike) {
 
 export function useSystemUpdate() {
   const { push } = useHistory();
-  const basePike = usePike('base');
+  const base = usePike('base');
   const update = base?.arak?.rail?.next?.[0];
-  const newKelvin = basePike?.wefts[0]?.kelvin ?? 418;
+  const newKelvin = base?.wefts[0]?.kelvin ?? 418;
   const charges = useCharges();
   const [blocked] = useKilnState((s) => {
     const [b, u] = partition(Object.entries(s.pikes), ([, pike]) => pikeIsBlocked(newKelvin, pike));
